+++ failgood/jvm/src/failgood/ExecutionListener.kt
@@ -4,4 +4,6 @@ internal interface ExecutionListener {
     suspend fun testStarted(testDescription: TestDescription)
     suspend fun testFinished(testPlusResult: TestPlusResult)
     suspend fun testEvent(testDescription: TestDescription, type: String, payload: String)
+    suspend fun testDiscovered(testDescription: TestDescription) {}
+    suspend fun contextDiscovered(context: Context) {}
 }