+++ failgood/jvm/src/failgood/Suite.kt
@@ -5,8 +5,6 @@ import failgood.internal.ContextInfo
 import failgood.internal.ContextTreeReporter
 import failgood.internal.ExecuteAllTestFilterProvider
 import failgood.internal.TestFilterProvider
-import failgood.internal.execution.context.ContextExecutor
-import failgood.internal.util.getenv
 import failgood.internal.util.pluralize
 import kotlinx.coroutines.*
 import java.lang.management.ManagementFactory
@@ -34,9 +32,6 @@ class Suite(val contextProviders: Collection<ContextProvider>) {
         }
     }
 
-    // set timeout to the timeout in milliseconds, an empty string to turn it off
-    private val timeoutMillis: Long = parseTimeout(getenv("TIMEOUT"))
-
     companion object {
         internal fun parseTimeout(timeout: String?): Long {
             return when (timeout) {
@@ -53,8 +48,11 @@ class Suite(val contextProviders: Collection<ContextProvider>) {
         executionFilter: TestFilterProvider = ExecuteAllTestFilterProvider,
         listener: ExecutionListener = NullExecutionListener
     ): List<Deferred<ContextResult>> {
-        val tag = getenv("FAILGOOD_TAG")
-        return contextProviders
+        return getRootContexts(coroutineScope).investigate(coroutineScope, executeTests, executionFilter, listener)
+    }
+
+    internal suspend fun getRootContexts(coroutineScope: CoroutineScope): LoadResults = LoadResults(
+        contextProviders
             .map {
                 coroutineScope.async {
                     try {
@@ -64,32 +62,7 @@ class Suite(val contextProviders: Collection<ContextProvider>) {
                     }
                 }
             }.flatMap { it.await() }.sortedBy { it.order }
-            .map { context: LoadResult ->
-                when (context) {
-                    is CouldNotLoadContext ->
-                        CompletableDeferred(
-                            FailedRootContext(Context(context.jClass.name ?: "unknown"), context.reason)
-                        )
-                    is RootContext -> {
-                        val testFilter = executionFilter.forClass(context.sourceInfo.className)
-                        coroutineScope.async {
-                            if (context.ignored?.isIgnored() == null) {
-                                ContextExecutor(
-                                    context,
-                                    coroutineScope,
-                                    !executeTests,
-                                    listener,
-                                    testFilter,
-                                    timeoutMillis,
-                                    runOnlyTag = tag
-                                ).execute()
-                            } else
-                                ContextInfo(emptyList(), mapOf(), setOf())
-                        }
-                    }
-                }
-            }
-    }
+    )
 }
 
 internal object NullExecutionListener : ExecutionListener {