+++ failgood/jvm/test/failgood/DescribeTest.kt
@@ -8,7 +8,7 @@ class DescribeTest {
     val context = describe("The describe top level method") {
         it("creates a context named '<className>' when called with a class") {
             expectThat(describe(String::class) {}) {
-                get { name }.isEqualTo("String")
+                get { context.name }.isEqualTo("String")
             }
         }
     }